found out about typedefs

+++ 02/02.rs
@@ -22,24 +22,24 @@ fn main() {
     println!("time: {:?}", start.elapsed());
 }
 
-fn part1(games: &[(char, char)]) -> i32 {
+fn part1(games: &[(char, char)]) -> usize {
     let diff = b'X' - b'A';
     let offset = b'W';
 
-    let mut score: i32 = 0;
+    let mut score: usize = 0;
     for &(opp, you) in games {
         if opp as u8 + diff as u8 == you as u8 {
             score += 3;
         } else if opp == 'A' && you == 'Y' || opp == 'B' && you == 'Z' || opp == 'C' && you == 'X' {
             score += 6;
         }
-        score += (you as u8 - offset as u8) as i32;
+        score += (you as u8 - offset as u8) as usize;
     }
     score
 }
 
-fn part2(games: &[(char, char)]) -> i32 {
-    let mut score: i32 = 0;
+fn part2(games: &[(char, char)]) -> usize {
+    let mut score: usize = 0;
     for &(opp, you) in games {
         if you == 'X' {
             // loss
@@ -52,7 +52,7 @@ fn part2(games: &[(char, char)]) -> i32 {
             }
         } else if you == 'Y' {
             // draw
-            score += (3 + (opp as u8 - b'A' + 1)) as i32;
+            score += (3 + (opp as u8 - b'A' + 1)) as usize;
         } else {
             // win
             score += 6;
